requirements.txt
absl-py==2.0.0
anndata==0.10.3
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.4
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
biopython==1.81
bleach==6.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chembl-webresource-client==0.10.8
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
easydict==1.11
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.44.3
fqdn==1.5.1
fsspec==2023.4.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
install==1.3.5
ipykernel==6.26.0
ipython==8.18.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
PubChemPy==1.0.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
rdkit-pypi==2022.9.5
referencing==0.31.0
requests==2.31.0
requests-cache==0.7.5
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
rsa==4.9
scanpy==1.9.6
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.12.2
Send2Trash==1.8.2
session-info==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spektral==1.3.0
stack-data==0.6.3
statsmodels==0.14.0
stdlib-list==0.10.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
torch==2.1.1+cu118
torch_geometric==2.4.0
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.5
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
